Split Tensor Dim Class from Tensor
Split Tensor Dim Class from Tensor for better usability

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped
