+++ debian/nntrainer-dev.install
@@ -5,6 +5,7 @@
 /usr/include/nntrainer/bn_layer.h
 /usr/include/nntrainer/neuralnet.h
 /usr/include/nntrainer/tensor.h
+/usr/include/nntrainer/tensor_dim.h
 /usr/include/nntrainer/nntrainer.h
 /usr/lib/*/pkgconfig/*.pc
 /usr/lib/*/*.a
