+++ nntrainer/src/tensor.cpp
@@ -36,35 +36,6 @@
 
 namespace nntrainer {
 
-int TensorDim::setTensorDim(std::string input_shape) {
-  int status = ML_ERROR_NONE;
-  std::regex words_regex("[^\\s.,:;!?]+");
-  auto words_begin =
-    std::sregex_iterator(input_shape.begin(), input_shape.end(), words_regex);
-  auto words_end = std::sregex_iterator();
-  int cur_dim = std::distance(words_begin, words_end);
-  if (cur_dim > 4) {
-    ml_loge("Tensor Dimension should be less than 4");
-    return ML_ERROR_INVALID_PARAMETER;
-  }
-  int cn = 0;
-  for (std::sregex_iterator i = words_begin; i != words_end; ++i) {
-    dim[MAXDIM - cur_dim + cn] = std::stoi((*i).str());
-    if (dim[MAXDIM - cur_dim + cn] <= 0) {
-      ml_loge("Tensor Dimension should be greater than 0");
-      return ML_ERROR_INVALID_PARAMETER;
-    }
-    cn++;
-  }
-  return status;
-}
-
-void TensorDim::operator=(const TensorDim &from){
-  for (int i=0; i< 4; ++i){
-    this->dim[i] = from.dim[i];
-  }
-};
-
 Tensor::Tensor(int height, int width) {
   this->height = height;
   this->width = width;
